pyproject.toml
[project]
name = "monitor-cli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "paho-mqtt>=2.1.0",
    "pyserial>=3.5",
    "rich>=14.2.0",
]
